RPM-To-Locked-Rotor: Initial commit
RPM-To-Locked-Rotor is a simple program designed to run on an ATtiny85
microcontroller. It accepts two RPM signal inputs from PC style fans
and drives two locked-rotor style outputs to emulate the behavior of
server/industrial-style fans.

RPM signal inputs are on pins 7 and 2. The associated locked-rotor
outputs are on pins 5 and 6. Both the inputs and outputs are intended
to be pulled up externally (outputs are open drain).

Signed-off-by: Jacob Moroni <mail@jakemoroni.com>